+++ src/main/java/com/paoloamosso/piggus/model/User.java
@@ -9,6 +9,7 @@
 
 package com.paoloamosso.piggus.model;
 
+import com.paoloamosso.piggus.converter.LocalDatePersistenceConverter;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
 import org.hibernate.annotations.GenericGenerator;
@@ -17,6 +18,7 @@ import org.hibernate.validator.constraints.Length;
 import javax.persistence.*;
 import javax.validation.constraints.Email;
 import javax.validation.constraints.NotEmpty;
+import java.time.LocalDate;
 import java.util.List;
 import java.util.Set;
 
@@ -34,6 +36,8 @@ public class User {
     private int idUser;
     @Column(name = "user_public_identifier")
     private String userPublicIdentifier;
+    @Column(name = "user_secret_identifier")
+    private String userSecretIdentifier;
     @Column(name = "username")
     @NotEmpty(message="*Please provide an username")
     private String username;
@@ -48,6 +52,12 @@ public class User {
     private int active;
     @Column(name = "recovery_mode")
     private int recoveryMode;
+    @Column(name = "registration_date")
+    @Convert(converter = LocalDatePersistenceConverter.class)
+    private LocalDate registrationDate;
+    @Column(name = "last_login")
+    @Convert(converter = LocalDatePersistenceConverter.class)
+    private LocalDate lastLogin;
     @Column(name = "accepted_privacy")
     private boolean acceptedPrivacy = false;
     @Column(name = "accepted_marketing")